src/main/java/com/neosoft/restapi/BookController.java
package com.neosoft.restapi;

import javassist.NotFoundException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping(value = "/book")
public class BookController {
    @Autowired
    BookRepository bookRepository;

    @GetMapping
    public List<Book> getAllBookRecords(){
        return bookRepository.findAll();
    }

    @GetMapping(value = "{bookId}")
    public Book getBookById(@PathVariable(value = "bookId") Long bookId){
        return bookRepository.findById(bookId).get();
    }

    @PostMapping
    public Book createBookRecord(@RequestBody @Valid Book bookRecord){
        return bookRepository.save(bookRecord);
    }

    @PutMapping
    public Book updateBookRecord(@RequestBody @Valid Book bookRecord) throws NotFoundException {
        if(bookRecord==null || bookRecord.getBookId()==null){
            throw new NotFoundException("BookRecord or ID must not be null!");
        }
        Optional<Book> optionalBook=bookRepository.findById(bookRecord.getBookId());
        if(!optionalBook.isPresent()){
            throw new NotFoundException("Book with Id: "+ bookRecord.getBookId() + " does not exist.");
        }
        Book existingBookRecord=optionalBook.get();
        existingBookRecord.setName(bookRecord.getName());
        existingBookRecord.setSummary(bookRecord.getSummary());
        existingBookRecord.setRating(bookRecord.getRating());
        return bookRepository.save(existingBookRecord);
    }

    @DeleteMapping(value = "/delete/{bookId}")
    public void deleteById(@PathVariable(value = "bookId") Long bookId) throws NotFoundException{
        if(!bookRepository.findById(bookId).isPresent()){
            throw new NotFoundException("book "+bookId+"already does not exist");
        }

        bookRepository.deleteById(bookId);
    }

}